# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chatbot Admin Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/client/login:
    post:
      operationId: clientLogin
      tags: [auth]
      summary: Client login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not authenticated
...
  /webhook/message:
    post:
      operationId: ingestChatMessage
      tags: [webhook]
      summary: Receive a chatbot conversation and store it as a chat log
      description: >
        Called by your chatbot after each conversation turn. Authenticate using
        the channel-specific API key configured in the company settings
        (telegramBotApiKey, whatsappApiToken, messengerApiKey, or websiteChatbotKey).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookMessageInput"
      responses:
        "201":
          description: Chat log saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookMessageResult"
        "401":
          description: Invalid API key
        "404":
          description: Company not found for the given key
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [role, name]
      properties:
        role:
          type: string
          enum: [admin, client]
        name:
          type: string
        id:
          type: ["integer", "null"]
    SessionUser:
      type: object
      required: [role, name]
      properties:
        role:
          type: string
          enum: [admin, client]
        name:
          type: string
        id:
          type: ["integer", "null"]
...
    WebhookMessageInput:
      type: object
      required: [apiKey, channel, customerMessage]
      properties:
        apiKey:
          type: string
          description: The channel-specific API key stored in company settings (telegramBotApiKey, whatsappApiToken, messengerApiKey, or websiteChatbotKey)
        channel:
          type: string
          enum: [telegram, whatsapp, messenger, website]
        sessionId:
          type: string
          description: Optional identifier for a customer conversation thread. Messages sharing the same sessionId are grouped as a conversation.
        customerMessage:
          type: string
        botResponse:
          type: ["string", "null"]
          description: The bot's reply to this message. Omit on the first call to retrieve context before generating a response.
    ConversationTurn:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    WebhookMessageResult:
      type: object
      required: [id, companyId, channel, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        channel:
          type: string
        sessionId:
          type: ["string", "null"]
        createdAt:
          type: string
        systemPrompt:
          type: ["string", "null"]
          description: The system prompt configured by the client — use this as the AI system message.
        conversationHistory:
          type: array
          description: Prior turns in this session (oldest first, excluding the current message). Feed into your AI call for multi-turn context.
          items:
            $ref: "#/components/schemas/ConversationTurn"